[#55] refactor : change interface brand location
